chore: rm old files, fix types on test
